update to latest foundation library

+++ network/tcp.c
@@ -513,7 +513,7 @@ static unsigned int _tcp_socket_buffer_read( socket_t* sock, unsigned int wanted
 				cols = ret - ( row * 8 );
 			for( ; col < cols; ++col, ofs +=3 )
 			{
-				string_format_buffer( dump_buffer + ofs, 66 - ofs, "%02x", *( src + ( row * 8 ) + col ) );
+				string_format_buffer( dump_buffer + ofs, 66 - (unsigned int)ofs, "%02x", *( src + ( row * 8 ) + col ) );
 				*( dump_buffer + ofs + 2 ) = ' ';
 			}
 			if( ofs )
@@ -596,7 +596,7 @@ static unsigned int _tcp_socket_buffer_write( socket_t* sock )
 					cols = res - ( row * 8 );
 				for( ; col < cols; ++col, ofs +=3 )
 				{
-					string_format_buffer( buffer + ofs, 34 - ofs, "%02x", *( src + ( row * 8 ) + col ) );
+					string_format_buffer( buffer + ofs, 34 - (unsigned int)ofs, "%02x", *( src + ( row * 8 ) + col ) );
 					*( buffer + ofs + 2 ) = ' ';
 				}
 				if( ofs )
